refactor: remove extra helper files
